feat: change the llama index retriever implementation to use FGAFilter
