+++ packages/llama-index-auth0-ai/llama_index_auth0_ai/fga/__init__.py
@@ -1,0 +1,3 @@
+from .FGARetriever import FGARetriever
+
+__all__ = ["FGARetriever"]
