+++ packages/llama-index-auth0-ai/pyproject.toml
@@ -14,6 +14,7 @@ readme = "README.md"
 python = "^3.11"
 llama-index = "^0.12.11"
 openfga-sdk = "^0.9.0"
+auth0-ai = {path = "../auth0-ai", develop = true}
 
 
 [tool.poetry.group.test.dependencies]
